refactor: consolidate audio cache into audioController, remove unused functions
- Remove unused clearAudioCache() and getAudioCacheStats() functions
- Move audio cache logic into audioController.js for single-file audio handling
- Delete src/utils/audioCache.js
